% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.tsets.estimate}
\alias{coef.tsets.estimate}
\alias{coef}
\title{Extract Model Coefficients}
\usage{
\method{coef}{tsets.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate}.}

\item{...}{not currently used.}
}
\value{
a numeric named vector.
}
\description{
Extract the estimated coefficients of a model.
}
