% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsets-package.R
\docType{package}
\name{tsets-package}
\alias{tsets}
\alias{tsets-package}
\title{tsets: Exponential Smoothing Models}
\description{
Exponential smoothing (ETS) models estimated using QML. Methods for estimation, prediction, simulation, inference and plotting.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

Other contributors:
\itemize{
  \item James Nesbit \email{jmcgnesbit@gmail.com} [contributor]
  \item Keith O'Hara \email{kth.ohr@gmail.com} [contributor]
}

}
\keyword{internal}
