% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{tsmoments.tsets.estimate}
\alias{tsmoments.tsets.estimate}
\alias{tsmoments}
\title{Analytic Forecast Moments}
\usage{
\method{tsmoments}{tsets.estimate}(object, h = 1, newxreg = NULL, init_states = NULL, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate}.}

\item{h}{the forecast horizon.}

\item{newxreg}{a matrix of external regressors in the forecast horizon.}

\item{init_states}{optional vector of states to initialize the forecast.
If NULL, will use the last available state from the estimated model.}

\item{...}{not currently used.}
}
\description{
Prediction function for class \dQuote{tsets.estimate}.
}
\note{
The function is currently incomplete as it does not provide an option
to backtransform the moments for Box Cox or logit transformed additive models.
}
