% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsreport.tsets.estimate}
\alias{tsreport.tsets.estimate}
\alias{tsreport}
\title{Estimation Summary Report (pdf format)}
\usage{
\method{tsreport}{tsets.estimate}(object, output_dir = "/", args = list(name = NULL), ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate}.}

\item{output_dir}{directory where the output is written and the object saved.}

\item{args}{additional list of arguments used in the generation of the report.
Only name is currently used to display the name of the data series.}

\item{...}{not currently used.}
}
\value{
A pdf file.
}
\description{
Generates a pdf summary of the estimated model.
}
