% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{auto_ets}
\alias{auto_ets}
\title{Automatic Model Selection}
\usage{
auto_ets(
  y,
  xreg = NULL,
  transformation = NULL,
  lambda = NULL,
  lower = 0,
  upper = 1,
  metric = "AIC",
  frequency = NULL,
  normalized_seasonality = TRUE,
  additive_only = FALSE,
  solver = "nlminb",
  control = list(trace = 0, maxit = 1000),
  power_model = FALSE,
  include_damped = TRUE,
  trace = FALSE,
  return_table = FALSE,
  scale = FALSE,
  seasonal_init = "fixed",
  autodiff = TRUE,
  ...
)
}
\arguments{
\item{y}{an xts vector.}

\item{xreg}{an optional xts matrix of regressors (pre-lagged).}

\item{transformation}{a valid transformation for y from the \dQuote{tstransform}
function in the \dQuote{tsaux} package (currently box-cox or logit are
available) applied to additive models only.}

\item{lambda}{the Box Cox power parameter (lambda). If NA will estimate this
using the method of Guerrero.}

\item{lower}{the lower bound for the transformation.}

\item{upper}{the upper bound for the transformation.}

\item{metric}{the selection metric to use. Valid metrics are \sQuote{AIC},
\sQuote{BIC}, \sQuote{AICc}, \sQuote{MASE} and \sQuote{MAPE}. If lambda is
not NULL, then those models which admit a Box-Cox transformation (additive
error models) will not be comparable with the other models.}

\item{frequency}{the frequency of y (if using a seasonal model).}

\item{normalized_seasonality}{whether to impose Roberts-McKenzie normalized
seasonality.}

\item{additive_only}{whether to limit to additive models only.}

\item{solver}{the solver to use for estimation.}

\item{control}{the solver control parameters.}

\item{power_model}{whether to include the power MAM models.}

\item{include_damped}{whether to include damped models in the selection.}

\item{trace}{whether to show the progress bar. The user is expected to have
set up appropriate handlers for this using the \dQuote{progressr} package.}

\item{return_table}{whether to return the table with the enumerated options,
ranked by metric,for each combination of those options used.}

\item{scale}{whether to rescale the data using y/max(y) (only for additive models).
This sometimes helps in the optimization.}

\item{seasonal_init}{whether the initial seasonal states are estimated or
fixed (set to a backcast approximation).}

\item{autodiff}{whether to use automatic differentiation
(see \code{\link{estimate.tsets.spec}}).}

\item{...}{not used.}
}
\value{
An object of class \dQuote{tsets.estimate} which also inherits
class \dQuote{tsets.select}
}
\description{
Automatic model selection based on choice of criteria.
}
\details{
The function estimates a number of different models from the class
of valid ETS models in the packages, with and without dampening, with and
without a power term (for MAM and MAN models).
}
\note{
The function can use parallel functionality as long as the user has set up a
\code{\link[future]{plan}} using the future package.
}
