% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.tsets.estimate}
\alias{fitted.tsets.estimate}
\alias{fitted}
\title{Model Fitted Values}
\usage{
\method{fitted}{tsets.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate}.}

\item{...}{not currently used.}
}
\description{
Extract the fitted values from an estimated model.
}
