% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{tsbenchmark.tsets.spec}
\alias{tsbenchmark.tsets.spec}
\alias{tsbenchmark}
\title{Model Benchmarking}
\usage{
\method{tsbenchmark}{tsets.spec}(
  object,
  solver = "nlminb",
  control = list(trace = 0),
  autodiff = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.spec}.}

\item{solver}{solver to use for estimation.}

\item{control}{additional controls passed to the solver.}

\item{autodiff}{whether to use automatic differentiation for estimation.
This makes use of the tsetsad package.}

\item{...}{not currently used.}
}
\description{
Benchmarks and saves the details of a model for benchmarking of
speed and unit testing.
}
