% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsdecompose.tsets.estimate}
\alias{tsdecompose.tsets.estimate}
\alias{tsdecompose}
\alias{tsdecompose.tsets.predict}
\title{Model Decomposition}
\usage{
\method{tsdecompose}{tsets.estimate}(object, simplify = FALSE, ...)

\method{tsdecompose}{tsets.predict}(object, simplify = FALSE, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate} or \dQuote{tsets.predict}}

\item{simplify}{whether to return the components as Trend (Level + Slope), Seasonal,
X and Irregular.}

\item{...}{not currently used.}
}
\value{
For the estimated object, returns an xts matrix of the state components
(including error). For the predicted object, a list with the simulated object
which includes the predictive distribution matrix of the state components and
their estimated values, inheriting class \dQuote{tsmodel.predict} (for pretty
plotting functionality).
}
\description{
Decomposes the estimated model or prediction into its component
parts (states).
}
