% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdiagnose.R
\name{tsdiagnose.tsets.estimate}
\alias{tsdiagnose.tsets.estimate}
\alias{tsdiagnose}
\title{Model Diagnostics}
\usage{
\method{tsdiagnose}{tsets.estimate}(object, plot = FALSE, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate}.}

\item{plot}{whether to generate diagnostic plots to accompany summary.}

\item{...}{not currently used.}
}
\value{
A list of tables (printed out and returned invisibly) with
Ljung-Box test for residual autocorrelation, parameter and model bounds
diagnostics and outlier dates using the Rosner test from the
\code{EnvStats} package.
}
\description{
Creates a short summary of model based diagnostics.
}
