% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmetrics.tsets.predict}
\alias{tsmetrics.tsets.predict}
\alias{tsmetrics}
\alias{tsmetrics.tsets.estimate}
\title{Performance Metrics}
\usage{
\method{tsmetrics}{tsets.predict}(object, actual, alpha = 0.1, ...)

\method{tsmetrics}{tsets.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsets.estimate} or \dQuote{tsets.predict}}

\item{actual}{the actual data matched to the dates of the forecasts.}

\item{alpha}{the coverage level for distributional forecast metrics.}

\item{...}{not currently used.}
}
\description{
Performance metrics from an estimated or predicted tsets model.
}
